New API implemented and:  1. Created a shared lib for linking the DFS apps.  2. Fixed dfs append bug.  3. Refactor in the client code
