+++ src/common/context_singleton.cc
@@ -1,5 +1,5 @@
 #include "context_singleton.hh"
 
+//Context* Context::singleton = nullptr;
 Context& context = *Context::connect();
 
-Context* Context::singleton = nullptr;
