+++ src/nodes/directory.cc
