+++ src/nodes/local_io.cc
@@ -24,8 +24,8 @@ void Local_io::write (std::string name, std::string& v) {
 // update {{{
 void Local_io::update (std::string name, std::string v, uint32_t pos, uint32_t len) {
   string file_path = disk_path + string("/") + name;
-  fstream file (file_path, ios::out | ios::binary);
-  file.seekp(pos, ios::beg);
+  fstream file (file_path, fstream::binary | fstream::in | fstream::out);
+  file.seekp(pos, ios_base::beg);
   file.write(v.c_str(), len);
   file.close();
 }
