+++ tests/metadata_test.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "messages/fileinfo.hh"
 #include "messages/blockinfo.hh"
-#include "fs/directory.hh"
+#include "nodes/directory.hh"
 #include "common/context_singleton.hh"
 
 using namespace eclipse;
